requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==3.0.4
Django==3.1.7
django-crispy-forms==1.11.2
django-filter==2.4.0
djangorestframework==3.12.4
idna==2.7
Markdown==3.3.4
python-restcountries==1.0.0
pytz==2021.1
requests==2.20.0
restcountries==1.1.3
six==1.11.0
sqlparse==0.4.1
urllib3==1.24.3
gunicorn
uWSGI>=2.0.18,<2.1
